# api/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce Website API
  description: |
    A comprehensive ecommerce API that provides functionality for user authentication, 
    product catalog management, shopping cart operations, order processing, and user profile management.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the JWT token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Response Format
    All API responses follow a consistent format:
    - Success responses include `success: true`, optional `message`, `data`, and `timestamp`
    - Error responses include `success: false`, `error` object with `code`, `message`, and optional `details`, and `timestamp`
  version: 1.0.0
  contact:
    name: API Support
    email: support@ecommerce.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.ecommerce.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/products:
    get:
      tags:
        - Products
      summary: Get products with filtering and pagination
      description: Retrieves a paginated list of products with optional filtering and sorting
      operationId: getProducts
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of items per page
        - name: category_id
          in: query
          schema:
            type: string
          description: Filter by category ID
        - name: min_price
          in: query
          schema:
            type: number
            minimum: 0
          description: Minimum price filter
        - name: max_price
          in: query
          schema:
            type: number
            minimum: 0
          description: Maximum price filter
        - name: in_stock
          in: query
          schema:
            type: boolean
          description: Filter by stock availability
        - name: search
          in: query
          schema:
            type: string
          description: Search term for product name or description
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [name, price, created_at, updated_at]
            default: created_at
          description: Field to sort by
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProductListResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: Get all categories
      description: Retrieves all product categories
      operationId: getCategories
      security: []
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          categories:
                            type: array
                            items:
                              $ref: '#/components/schemas/Category'
                          total:
                            type: integer
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Categories
      summary: Create a new category
      description: Creates a new product category (admin only)
      operationId: createCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/CategoryExistsError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data (varies by endpoint)
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid request data"
        details:
          type: object
          description: Additional error details

    # Authentication Schemas
...
    Product:
      type: object
      required:
        - id
        - name
        - price
        - sku
        - inventory
        - isActive
        - categoryId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Wireless Headphones"
        description:
          type: string
          example: "High-quality wireless headphones with noise cancellation"
        price:
          type: number
          format: float
          minimum: 0
          example: 199.99
        compareAtPrice:
          type: number
          format: float
          minimum: 0
          nullable: true
          example: 249.99
        sku:
          type: string
          example: "WH-001"
        inventory:
          type: integer
          minimum: 0
          example: 50
        isActive:
          type: boolean
          example: true
        categoryId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        images:
          type: array
          items:
            type: string
          example: ["https://example.com/image1.jpg", "https://example.com/image2.jpg"]
        specifications:
          type: object
          additionalProperties: true
          example:
            color: "Black"
            weight: "250g"
            battery: "30 hours"
        seoTitle:
          type: string
          nullable: true
          example: "Best Wireless Headphones - Premium Audio"
        seoDescription:
          type: string
          nullable: true
          example: "Experience premium audio quality with our wireless headphones"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        category:
          $ref: '#/components/schemas/Category'
    ProductListResponse:
      type: object
      required:
        - products
        - pagination
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    PaginationInfo:
      type: object
      required:
        - page
        - pageSize
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        pageSize:
          type: integer
          minimum: 1
          example: 20
        total:
          type: integer
          minimum: 0
          example: 100
        totalPages:
          type: integer
          minimum: 0
          example: 5

    # Category Schemas
    Category:
      type: object
      required:
        - id
        - name
        - slug
        - isActive
        - sortOrder
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Electronics"
        slug:
          type: string
          example: "electronics"
        description:
          type: string
          nullable: true
          example: "Electronic devices and accessories"
        parentId:
          type: string
          format: uuid
          nullable: true
          example: "123e4567-e89b-12d3-a456-426614174000"
        isActive:
          type: boolean
          example: true
        sortOrder:
          type: integer
          example: 0
        createdAt:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-01T12:00:00Z"
        children:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        parent:
          $ref: '#/components/schemas/Category'
    CreateCategoryRequest:
      type: object
      required:
        - name
        - slug
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Electronics"
        slug:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-z0-9-]+$'
          example: "electronics"
        description:
          type: string
          maxLength: 500
          example: "Electronic devices and accessories"
        parentId:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        isActive:
          type: boolean
          default: true
          example: true
        sortOrder:
          type: integer
          default: 0
          example: 0

    # Order Schemas
...
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid request data"
              details: "Email is required"
            timestamp: "2024-01-01T12:00:00Z"
    UnauthorizedError:
      description: Unauthorized access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
            timestamp: "2024-01-01T12:00:00Z"
    ForbiddenError:
      description: Forbidden access
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INSUFFICIENT_PERMISSIONS"
              message: "Access denied: insufficient permissions"
            timestamp: "2024-01-01T12:00:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
            timestamp: "2024-01-01T12:00:00Z"

    # Specific Error Responses
...
    CategoryExistsError:
      description: Category already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: "CATEGORY_EXISTS"
              message: "Category with this slug already exists"
            timestamp: "2024-01-01T12:00:00Z"
...